+++ lib/GoogleSheetUtils.ts
@@ -1,0 +1,68 @@
+const fs = require('fs').promises;
+import path from 'path';
+import process from 'process';
+import { authenticate } from '@google-cloud/local-auth';
+import { google } from 'googleapis';
+
+const SCOPES = ['https://www.googleapis.com/auth/spreadsheets'];
+const TOKEN_PATH = path.join(process.cwd(), 'token.json');
+const CREDENTIALS_PATH = path.join(process.cwd(), '/test-data/credentials.json');
+
+export class GoogleSheetUtils {
+
+  async loadSavedCredentialsIfExist() {
+    try {
+      const content = await fs.readFile(TOKEN_PATH);
+      const credentials = JSON.parse(content);
+      return google.auth.fromJSON(credentials);
+    } catch (err) {
+      return null;
+    }
+  }
+
+  async saveCredentials(client) {
+    const content = await fs.readFile(CREDENTIALS_PATH);
+    const keys = JSON.parse(content);
+    const key = keys.installed || keys.web;
+    const payload = JSON.stringify({
+      type: 'authorized_user',
+      client_id: key.client_id,
+      client_secret: key.client_secret,
+      refresh_token: client.credentials.refresh_token,
+    });
+    await fs.writeFile(TOKEN_PATH, payload);
+  }
+
+  async authorize() {
+    let client = await this.loadSavedCredentialsIfExist();
+    if (client) {
+      return client;
+    }
+    // @ts-ignore
+    client = await authenticate({
+      scopes: SCOPES,
+      keyfilePath: CREDENTIALS_PATH,
+    });
+    if (client.credentials) {
+      await this.saveCredentials(client);
+    }
+    return client;
+  }
+
+  async readExcel() {
+    let auth = await this.authorize();
+
+    const sheets = google.sheets({ version: 'v4', auth });
+    const res = await sheets.spreadsheets.values.get({
+      spreadsheetId: '1lmT9zWb7zwYP-C2GqTIgOAgNLLhw1oXcBpI6nj4jsN8',
+      range: 'Sheet1',
+    });
+    const rows = res.data.values;
+    if (!rows || rows.length === 0) {
+      console.log('No data found.');
+      return;
+    }
+
+    return rows;
+  }
+}
